allow to ignore errors when updating objects fails (#6)
This is useful when previously legal objects are not legal anymore,
without being updated. For example, as a result of an admission
control change. Since a human is needed to bring the old object
into conformance with the new requirements, the controller won't
be able to restart these objects. In such a scenario, it should
still be able to continue restarting other deployments until the
affected one is fixed.